fix paths

+++ cron.sh
@@ -1,6 +1,6 @@
-# run every 10 minutes
-# map files to webdirectory
-
-git pull
-docker-compose up --build
-docker image prune -a --filter "until=24h"
+if [[ -d "/mnt/backup/logs/fl" ]]
+then
+	git pull
+	docker-compose up --build
+	docker image prune -a --filter "until=24h"
+fi
